improved performance, now working on adding swipe tabs for allsongs and playlists
