+++ app/src/main/AndroidManifest.xml
@@ -1,14 +1,16 @@
 <?xml version="1.0" encoding="utf-8"?>
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
-    package="com.rahulk11.audioplayer">
+    package="com.rahulk11.randomplayer"
+    android:versionCode="110001"
+    android:versionName="1.0">
     <uses-permission android:name="android.permission.READ_EXTERNAL_STORAGE" />
     <uses-permission android:name="android.permission.READ_PHONE_STATE"/>
     <uses-permission android:name="android.permission.MODIFY_AUDIO_SETTINGS"/>
     <application
         android:allowBackup="true"
-        android:icon="@mipmap/ic_launcher"
+        android:icon="@drawable/music"
         android:label="@string/app_name"
-        android:roundIcon="@mipmap/ic_launcher_round"
+        android:roundIcon="@drawable/music"
         android:supportsRtl="true"
         android:theme="@style/AppTheme">
         <activity android:name=".MainActivity"
@@ -21,12 +23,12 @@
             </intent-filter>
         </activity>
         <service android:name=".SongService"/>
-        <receiver android:name="com.rahulk11.audioplayer.NotificationHandler$NotifBtnClickReceiver" >
+        <receiver android:name="com.rahulk11.randomplayer.helpers.NotificationHandler$NotifBtnClickReceiver" >
             <intent-filter>
-                <action android:name="com.rahulk11.audioplayer.ACTION_PLAY" />
-                <action android:name="com.rahulk11.audioplayer.ACTION_NEXT" />
-                <action android:name="com.rahulk11.audioplayer.ACTION_PREV" />
-                <action android:name="com.rahulk11.audioplayer.ACTION_CLOSE" />
+                <action android:name="com.rahulk11.randomplayer.ACTION_PLAY" />
+                <action android:name="com.rahulk11.randomplayer.ACTION_NEXT" />
+                <action android:name="com.rahulk11.randomplayer.ACTION_PREV" />
+                <action android:name="com.rahulk11.randomplayer.ACTION_CLOSE" />
                 <action android:name="android.intent.action.HEADSET_PLUG" />
             </intent-filter>
         </receiver>
