+++ app/src/main/java/com/rahulk11/randomplayer/AllSongs.java
@@ -1,0 +1,21 @@
+package com.rahulk11.randomplayer;
+
+import android.support.v4.app.Fragment;
+import android.os.Bundle;
+import android.support.annotation.Nullable;
+import android.view.LayoutInflater;
+import android.view.View;
+import android.view.ViewGroup;
+
+/**
+ * Created by rahul on 6/22/2017.
+ */
+
+public class AllSongs extends Fragment {
+    @Nullable
+    @Override
+    public View onCreateView(LayoutInflater inflater, @Nullable ViewGroup container, Bundle savedInstanceState) {
+        View v = inflater.inflate(R.layout.allsongs, container, false);
+        return v;
+    }
+}
