+++ app/src/main/java/com/rahulk11/randomplayer/SongService.java
@@ -1,24 +1,22 @@
-package com.rahulk11.audioplayer;
+package com.rahulk11.randomplayer;
 
 import android.app.Service;
-import android.content.ComponentName;
 import android.content.Context;
 import android.content.Intent;
 import android.content.IntentFilter;
-import android.graphics.Bitmap;
-import android.graphics.BitmapFactory;
 import android.media.AudioManager;
 import android.media.MediaMetadataRetriever;
 import android.media.MediaPlayer;
-import android.os.Handler;
 import android.os.IBinder;
 import android.support.annotation.Nullable;
 import android.telephony.PhoneStateListener;
 import android.telephony.TelephonyManager;
 import android.util.Log;
 
+import com.rahulk11.randomplayer.helpers.NotificationHandler;
+import com.rahulk11.randomplayer.helpers.PlaybackManager;
+
 import java.io.IOException;
-import java.io.InterruptedIOException;
 
 /**
  * Created by rahul on 6/9/2017.
@@ -220,12 +218,13 @@ public class SongService extends Service {
                         title = intent.getStringExtra(MainActivity.SONG_TITLE);
                         artist = intent.getStringExtra(MainActivity.ARTIST_NAME);
                         album = intent.getStringExtra(MainActivity.ALBUM_NAME);
+                        boolean resume = intent.getBooleanExtra("resume", false);
                         if (player != null) {
                             try {
                                 result = audioManager.requestAudioFocus(focusChangeListener, AudioManager.STREAM_MUSIC,
                                         AudioManager.AUDIOFOCUS_GAIN);
                                 if (result == AudioManager.AUDIOFOCUS_REQUEST_GRANTED) {
-                                    if (!android.text.TextUtils.isEmpty(data)) {
+                                    if (!android.text.TextUtils.isEmpty(data) && resume) {
                                         player.reset();
                                         player.setDataSource(data);
                                         player.prepare();
@@ -234,10 +233,10 @@ public class SongService extends Service {
                                             mmr.setDataSource(data);
                                             byteData = mmr.getEmbeddedPicture();
                                         }
-                                        player.seekTo(seekTo);
-                                        player.start();
-                                        notificationHandler.showNotif(byteData, title, artist, album, true);
                                     }
+                                    player.seekTo(seekTo);
+                                    player.start();
+                                    notificationHandler.showNotif(byteData, title, artist, album, true);
                                 }
                                 PlaybackManager.goAhead = true;
                             } catch (IOException e) {
