+++ app/src/main/res/layout/notif_new_layout.xml
@@ -8,7 +8,7 @@
 
     <FrameLayout
         android:layout_width="0dp"
-        android:layout_height="wrap_content"
+        android:layout_height="match_parent"
         android:layout_weight="1.5">
 
         <ImageView
@@ -124,7 +124,7 @@
     <FrameLayout
         android:id="@+id/fl_albumArt"
         android:layout_width="0dp"
-        android:layout_height="wrap_content"
+        android:layout_height="match_parent"
         android:layout_weight="0.50">
 
         <ImageView
