+++ app/src/main/res/layout/slidingup_panellayout.xml
@@ -5,7 +5,7 @@
     android:background="#ffffff"
     android:orientation="vertical">
 
-    <com.rahulk11.audioplayer.slidinguppanelhelper.SlidingUpPanelLayout xmlns:rahulk11="http://schemas.android.com/apk/res-auto"
+    <com.rahulk11.randomplayer.slidinguppanelhelper.SlidingUpPanelLayout xmlns:rahulk11="http://schemas.android.com/apk/res-auto"
         android:id="@+id/sliding_layout"
         android:layout_width="match_parent"
         android:layout_height="match_parent"
@@ -31,6 +31,6 @@
             layout="@layout/slidingpanel_childtwo"
             android:layout_width="match_parent"
             android:layout_height="match_parent" />
-    </com.rahulk11.audioplayer.slidinguppanelhelper.SlidingUpPanelLayout>
+    </com.rahulk11.randomplayer.slidinguppanelhelper.SlidingUpPanelLayout>
 
 </LinearLayout>
