+++ app/src/main/res/values/colors.xml
@@ -4,4 +4,5 @@
     <color name="colorPrimaryDark">#303F9F</color>
     <color name="colorAccent">#FF4081</color>
     <color name="colorTransparent">#00000000</color>
+    <color name="tabsScrollColor">#8a140e</color>
 </resources>
